% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_annotated_clustering}
\alias{add_annotated_clustering}
\title{add_annotated_clustering}
\usage{
add_annotated_clustering(loom, group, name, clusters, annotation,
  is.default = F)
}
\arguments{
\item{loom}{The loom file handler.}

\item{group}{The for the given clustering group to which the given clusters have to be added.}

\item{name}{The name given to this clustering}

\item{clusters}{A named list of the cell id and assigned the cluster id.}

\item{annotation}{A named list of the cell id and the corresponding annotation.}

\item{is.default}{Set this clustering be set as default one.}
}
\description{
Add the given clusters in the given group column attribute and meta data related to the given clustering to the given .loom file handler.
}
