% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_annotation_by_cluster_annotation_mapping_df}
\alias{add_annotation_by_cluster_annotation_mapping_df}
\title{add_annotation_by_cluster_annotation_mapping_df}
\usage{
add_annotation_by_cluster_annotation_mapping_df(loom, key, clustering.name,
  annotation.df = NULL, annotation.df.cluster.id.column.name = NULL,
  annotation.df.annotation.column.name = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of the annotation}

\item{clustering.name}{The name of the clustering.}

\item{annotation.df}{The data.frame containing the cluster annotatiion mapping}

\item{annotation.df.cluster.id.column.name}{The column name of the given annotation.df containing the cluster ID values}

\item{annotation.df.annotation.column.name}{The column name of the given annotation.df containing the annotation values.}
}
\description{
Add annotation as column attribute using cluster annotation mapping data.frame
}
