% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_clustering2}
\alias{add_clustering2}
\title{add_clustering2}
\usage{
add_clustering2(loom, group, name, clusters, level = 0, params = NULL,
  parent.clustering.id = NULL, annotation.df = NULL,
  annotation.df.cluster.id.cl = NULL,
  annotation.df.cluster.description.cl = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{group}{The for the given clustering group to which the given clusters have to be added}

\item{name}{The name given to this clustering}

\item{clusters}{A named list of the cell id and assigned the cluster id.}

\item{parent.clustering.id}{Clustering ID of the parent in case it's a subclustering.}

\item{is.default}{Set this clustering be set as default one.}

\item{annotation}{A data.frame with annotation for the clusters}

\item{annotation.cluster.id.cl}{The column name to use for the IDs of the clusters found by the given clustering group.}

\item{annotation.cluster.description.cl}{The column name to use for the description of the clusters found by the given clustering group.}
}
\description{
Add the given clusters in the given group column attribute and meta data related to the given clustering to the given .loom file handler.
}
