% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_col_attr}
\alias{add_col_attr}
\title{add_col_attr}
\usage{
add_col_attr(loom, key, value, dtype = NULL, as.annotation = F,
  as.metric = F)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of the new added attribute.}

\item{value}{The value of the new added attribute.}

\item{as.annotation}{Define this attribute as a discrete attribute. This attribute will be visible in the compare tab.}

\item{as.metric}{Define this attribute as a continues}
}
\description{
Add a new column attribute to the given .loom object accessible by the given key and containing the given value.
}
