% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_default_embedding}
\alias{add_default_embedding}
\title{add_default_embedding}
\usage{
add_default_embedding(loom, embedding)
}
\arguments{
\item{loom}{The loom file handler.}

\item{embedding}{The embedding to add.}
}
\description{
Add the main (default) embedding to the given loom file handler.
}
