% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_embedding}
\alias{add_embedding}
\title{add_embedding}
\usage{
add_embedding(loom, embedding, name, is.default = F,
  parent.clustering.id = NULL, parent.clustering.id.lookup.query = NULL,
  cluster.id = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{embedding}{A M-by-2 data.frame of the embeddings with M cells.}

\item{name}{The name of the given embedding.}

\item{is.default}{Default embedding to use in the .loom file.}
}
\description{
Add the given embedding as a row attribute and meta data related to the given embeddding to the given .loom file handler.
}
