% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_embedding_from_dyno}
\alias{add_embedding_from_dyno}
\title{add_embedding_dyno}
\usage{
add_embedding_from_dyno(loom, dyno, name = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{dyno}{A dyno object generated by dyno::infer_trajectory.}

\item{name}{A name to give to the embedding/trajectory inferred by the trajectory method.}
}
\description{
Add the given embedding as a column attribute and meta data related to the given embeddding to the given .loom file handler.
}
