% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_fbgn}
\alias{add_fbgn}
\title{add_fbgn}
\usage{
add_fbgn(loom, dgem, fbgn.gn.mapping.file.path)
}
\arguments{
\item{loom}{The loom file handler.}

\item{dgem}{A matrix of the gene expression with M genes as rows and N cells as columns.}

\item{fbgn.gn.mapping.file.path}{A N-by-2 data.frame containing the mapping between the Flybase gene and the gene symbol.}
}
\description{
Add the Flybase gene as a row attribute to the given .loom file handler.
}
