% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R, R/loom_mine.R
\name{add_global_attr}
\alias{add_global_attr}
\alias{add_global_attr}
\title{add_global_attr}
\usage{
add_global_attr(loom, key, value)

add_global_attr(loom, key, value)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of the new added attribute.}

\item{value}{The value of the new added attribute.}

\item{loom}{The loom file handler.}

\item{key}{The name of the new added attribute.}

\item{value}{The value of the new added attribute.}
}
\description{
Add a new global attribute to the given .loom object accessible by the given key and containing the given value.

Add a new global attribute to the given .loom object accessible by the given key and containing the given value.
}
