% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R, R/loom_mine.R
\name{add_global_md_annotation}
\alias{add_global_md_annotation}
\alias{add_global_md_annotation}
\title{add_global_md_annotation}
\usage{
add_global_md_annotation(loom, name, values)

add_global_md_annotation(loom, name, values)
}
\arguments{
\item{loom}{The loom file handler.}

\item{name}{The name of the annotation.}

\item{values}{The values of the annotation to be added.}

\item{loom}{The loom file handler.}

\item{name}{The name of the annotation.}

\item{values}{The values of the annotation to be added.}
}
\description{
Add the embedding with the given name to the global MetaData attribute.

Add the embedding with the given name to the global MetaData attribute.
}
