% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_global_md_clustering_kv}
\alias{add_global_md_clustering_kv}
\title{add_global_md_clustering_kv}
\usage{
add_global_md_clustering_kv(loom, clustering.id, key, value)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.id}{The clustering ID of the clustering which the key/value is added to.}

\item{key}{The name of the attribute to add.}

\item{value}{The value of the attribute to add.}
}
\description{
Add given value with the given key to the clustering with the given clustering.id of the global MetaData attribute
}
