% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_global_md_metric}
\alias{add_global_md_metric}
\title{add_global_md_metric}
\usage{
add_global_md_metric(loom, name)
}
\arguments{
\item{loom}{The loom file handler.}

\item{name}{The name of the annotation.}

\item{values}{The values of the annotation to be added.}
}
\description{
Add the metric with the given name to the global MetaData attribute.
}
