% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_hierarchy}
\alias{add_hierarchy}
\title{add_hierarchy}
\usage{
add_hierarchy(loom, hierarchy)
}
\arguments{
\item{loom}{The loom file handler.}

\item{hierarchy}{A named list of the 3-levels names. Use create_hierarchy() to build it.}
}
\description{
Add a 3-level hierarchy as global attributes to the given loom.
}
