% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_matrix}
\alias{add_matrix}
\title{add_matrix}
\usage{
add_matrix(loom, dgem, chunk.size, display.progress)
}
\arguments{
\item{loom}{The loom file handler.}

\item{dgem}{A matrix of the gene expression with M genes as rows and N cells as columns.}
}
\description{
Add the given gene expression matrix dgem to the given .loom object.
}
