% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_regulons_AUC}
\alias{add_regulons_AUC}
\title{add_regulons_AUC}
\usage{
add_regulons_AUC(loom, regulons.AUC, column.attr.name = "MotifRegulonsAUC")
}
\arguments{
\item{loom}{The loom file handler.}

\item{regulons.AUC}{A matrix of the regulons AUC values with M regulons as rows and N cells as columns.}

\item{column.attr.name}{Type of regulon to retrieve; Usually 'MotifRegulonsAUC' for motif-based regulons, or 'TrackRegulonsAUC' for track-based (e.g. ChIP-seq) regulons. 
(Might change according to the SCENIC/pySCENIC pipeline version and settings used).}
}
\description{
Add the regulons AUC matrix generated by SCENIC as a column attribute to the given .loom file handler.
}
