% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_scenic_regulons}
\alias{add_scenic_regulons}
\title{add_scenic_regulons}
\usage{
add_scenic_regulons(
  loom,
  regulons,
  regulon.threshold.assignments = NULL,
  regulon.enrichment.table = NULL
)
}
\arguments{
\item{loom}{The loom file handler.}

\item{regulons}{A list of list of the regulons and their target genes generated by SCENIC.}

\item{regulon.threshold.assignments}{The automated regulon cell assignments object generated by SCENIC.}

\item{regulon.enrichment.table}{The regulon enrichment table generated by SCENIC.}
}
\description{
Add the regulons with their target genes generated by SCENIC as a row attribute to the given .loom file handler.
}
