% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{append_clustering_update_ca}
\alias{append_clustering_update_ca}
\title{append_clustering_update_ca}
\usage{
append_clustering_update_ca(loom, clustering.id, clustering)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.id}{The id of the embedding group.}

\item{clustering}{The embedding to add.}
}
\description{
Append the given clustering with the given clustering.id.
}
