% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{check_global_meta_data}
\alias{check_global_meta_data}
\title{check_global_meta_data}
\usage{
check_global_meta_data(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Perform some sanity checks on the MetaData global attribute
}
