% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{check_missings}
\alias{check_missings}
\title{check_missings}
\usage{
check_missings(x, y, check.both = TRUE)
}
\arguments{
\item{x}{A vector of elements}

\item{y}{Another vector of elements}
}
\value{
A list with "result", a logical whether the two vectors are conserved
                   "x.missing.in.y", a vector of elements in x that are missing in y
                   "y.missing.in.x", a vector of elements in y that are missing in x
}
\description{
Check if some elements are missing between given vectors x an y
}
