% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{chunk_points}
\alias{chunk_points}
\title{chunk_points}
\usage{
chunk_points(data.size, chunk.size)
}
\arguments{
\item{data.size}{How big is the data being chunked}

\item{chunk.size}{How big should each chunk be}
}
\value{
A matrix where each column is a chunk, row 1 is start points, row 2 is end points
}
\description{
Generate chunk points
}
