% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{clear_global_meta_data}
\alias{clear_global_meta_data}
\title{clear_global_meta_data}
\usage{
clear_global_meta_data(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Empty the global MetaData attribute.
}
