% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{close_loom}
\alias{close_loom}
\title{close_loom}
\usage{
close_loom(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Close the given .loom file handler. Equivalent to finalize but without flushing.
}
