% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{col_attrs_exists_by_key}
\alias{col_attrs_exists_by_key}
\title{col_attrs_exists_by_key}
\usage{
col_attrs_exists_by_key(loom, key)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of the new added attribute.}
}
\description{
Check if the column attribute with the given key exists.
}
