% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{compress_gzb64}
\alias{compress_gzb64}
\title{open_loom}
\usage{
compress_gzb64(c)
}
\arguments{
\item{file.path}{The file path to the .loom file.}
}
\description{
Open loom file and return a .loom file handler.
}
