% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{convert_to_loom_v3_spec}
\alias{convert_to_loom_v3_spec}
\title{convert_to_loom_v3_spec}
\usage{
convert_to_loom_v3_spec(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\value{
A loom file handler
}
\description{
Convert a .loom file to version 3 specification of the Loom format
}
