% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create_cluster_annotation}
\alias{create_cluster_annotation}
\title{create_cluster_annotation}
\usage{
create_cluster_annotation(clusters, cluster.meta.data.df = NULL,
  cluster.id.cn = NULL, cluster.description.cn = NULL)
}
\arguments{
\item{clusters}{A named list of the cell id and the assigned cluster id generated by a clustering method (e.g.: Seurat).}

\item{cluster.meta.data.df}{A data.frame object with at least 2 columns named cluster.id.cn and cluster.description.cn.}

\item{cluster.id.cn}{The column name to use for the IDs of the clusters.}

\item{cluster.description.cn}{The column name to use for the description of the clusters.}
}
\description{
Create a named list where names are the cell IDs and the values are the annotation stored in the given cluster.description.cn column of the given cluster.meta.data.df. 
            If cluster.meta.data.df is set to NULL, annotation of cells are set to "NDA - Cluster X" where X is the cluster ID.
}
