% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create_hierarchy}
\alias{create_hierarchy}
\title{create_hierarchy}
\usage{
create_hierarchy(level.1.name = NULL, level.2.name = NULL,
  level.3.name = NULL)
}
\arguments{
\item{level.1.name}{Name of top level of the hierarchy.}

\item{level.2.name}{Name of the second level of the hierarchy.}

\item{level.3.name}{Name of the third level of the hierarchy.}
}
\description{
Create hierarchy as named list of tree levels (SCopeTreeL1, SCopeTreeL2, SCopeTreeL3)
}
