% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create_trajectory}
\alias{create_trajectory}
\title{create_trajectory}
\usage{
create_trajectory(edges, coordinates)
}
\arguments{
\item{cds}{CellDataSet object from monocle.}
}
\description{
Create a trajectory object from the given edges and coordinates of the nodes.
}
