% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create_trajectory_from_dyno}
\alias{create_trajectory_from_dyno}
\title{create_trajectory_from_dyno}
\usage{
create_trajectory_from_dyno(dyno)
}
\arguments{
\item{dyno}{A dyno object generated by dyno::infer_trajectory.}
}
\description{
Create a trajectory object from dyno TI output.
}
