% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create_trajectory_from_monocle}
\alias{create_trajectory_from_monocle}
\title{create_trajectory_from_monocle}
\usage{
create_trajectory_from_monocle(cds)
}
\arguments{
\item{cds}{CellDataSet object from monocle.}
}
\description{
Create a trajectory object from CellDataSet monocle output.
}
