% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{decompress_gzb64}
\alias{decompress_gzb64}
\title{open_loom}
\usage{
decompress_gzb64(gzb64c)
}
\arguments{
\item{file.path}{The file path to the .loom file.}
}
\description{
Open loom file and return a .loom file handler.
}
