% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_cell_annotation}
\alias{get_cell_annotation}
\title{get_cell_annotation}
\usage{
get_cell_annotation(loom, annotations.columns = NULL)
}
\arguments{
\item{loom}{.loom file name}
}
\value{
data.frame containing the cell annotations in the given loom (cells as rows, annotation variables as columns).
}
\description{
Get cell annotations from the given loom file
}
\examples{
loomPath <-  file.path(system.file('extdata', package='SCopeLoomR'), "example.loom")
loom <- open_loom(loomPath)

cell_annotations <- get_cell_annotation(loom)
head(cellAnnots)
}
