% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_cell_mask_by_cluster_name}
\alias{get_cell_mask_by_cluster_name}
\title{get_cell_mask_by_cluster_name}
\usage{
get_cell_mask_by_cluster_name(loom, cluster.name)
}
\arguments{
\item{loom}{The loom file handler.}

\item{cluster.name}{The name of the cluster.}
}
\value{
A N-by-1 boolean vector specifying which cells belong to the given cluster.name in the given loom.
}
\description{
Get a cell mask for the given cluster.name of the given loom.
}
