% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_cluster_info_by_id}
\alias{get_cluster_info_by_id}
\title{get_cluster_info_by_id}
\usage{
get_cluster_info_by_id(loom, clustering.id, cluster.idx, field = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{field}{The name of the field you want to get the value from.}

\item{clustering_id}{The ID of the clustering you want to get the field value from.}

\item{cluster_id}{`The ID of the cluster you want to get the field value from.}
}
\description{
Get the value of the given field from the cluster with the given clustering_idx and cluster_idx
}
