% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_cluster_markers}
\alias{get_cluster_markers}
\title{get_cluster_markers}
\usage{
get_cluster_markers(
  loom,
  clustering.id = NULL,
  clustering.name = NULL,
  log.fc.threshold = 1.5,
  adj.p.value.threshold = 0.05,
  n.signif.digits = NULL
)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.name}{The name/description of the clustering.}

\item{log.fc.threshold}{The log fold change threshold}

\item{adj.p.value.threshold}{The adjusted p-value threshold}
}
\value{
A list of L data.frame of N-by-M containing the statistical markers for all the clusters of from the clusteirng with the given clustering.name. L represents the number of clusters. N represents the number of statistical markers for a given cluster.
}
\description{
Get a list of data.frame each containing the statistical markers for the clusters from the clustering with the given clustering.name and filtered with the given log.fc.threshold and given adj.p.value.
}
