% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_cluster_names}
\alias{get_cluster_names}
\title{get_cluster_names}
\usage{
get_cluster_names(clustering)
}
\arguments{
\item{clustering}{The clustering object from the global MetaData attribute}

\item{loom}{The loom file handler.}
}
\description{
Get the cluster names of the given clustering object
}
