% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_clustering_by_id}
\alias{get_clustering_by_id}
\title{get_global_md_clustering_by_id}
\usage{
get_clustering_by_id(loom, clustering.id)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.id}{The ID of the clustering.}
}
\value{
A N-by-1 vector containing the cell assignments to each of the clusters of the clustering.
}
\description{
Get clustering with the given clustering.id of the given loom.
}
