% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_clustering_idx_by_cluster_name}
\alias{get_clustering_idx_by_cluster_name}
\title{get_clustering_idx_by_cluster_name}
\usage{
get_clustering_idx_by_cluster_name(loom, cluster.name)
}
\arguments{
\item{loom}{The loom file handler.}

\item{cluster.name}{The name of the cluster.}
}
\value{
The index of the clustering in the clusterings metadata global attribute corresponding to the clustering where the given cluster.name is found.
}
\description{
Get index of the clutering related to the given cluster.name.
}
