% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_clustering_markers_by_clustering_name}
\alias{get_clustering_markers_by_clustering_name}
\title{get_clustering_markers_by_clustering_name}
\usage{
get_clustering_markers_by_clustering_name(
  loom,
  clustering.name,
  log.fc.threshold = 1.5,
  adj.p.value = 0.05
)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.name}{The name/description of the clustering.}

\item{log.fc.threshold}{The log fold change threshold}

\item{adj.p.value}{The adjusted p-value threshold}
}
\value{
A list of L data.frame of N-by-M containing the statistical markers for all the clusters of from the clusteirng with the given clustering.name. L represents the number of clusters. N represents the number of statistical markers for a given cluster.
}
\description{
Get a list of data.frame each containing the statistical markers for the clusters from the clustering with the given clustering.name and filtered with the given log.fc.threshold and given adj.p.value.
}
