% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_clusterings}
\alias{get_clusterings}
\title{get_clusterings}
\usage{
get_clusterings(loom)
}
\arguments{
\item{loom}{The loom file handler.}

\item{cluster.name}{The name of the cluster.}
}
\value{
A N-by-M data.frame containing the clusterings of the given loom. N represents the cells and M the clusterings.
}
\description{
Get clustering of the given loom.
}
