% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_clusterings_with_name}
\alias{get_clusterings_with_name}
\title{get_clusterings_with_name}
\usage{
get_clusterings_with_name(loom)
}
\arguments{
\item{loom}{.loom file name}
}
\value{
data.frame containing the clusterings in the given loom (cells as rows, clustering alternatives as columns).
}
\description{
Get cell clusterings stored in the given loom file
}
\examples{
# clusters <- get_clusterings_with_name(loom)
# head(clusters)
}
