% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_default_embedding}
\alias{get_default_embedding}
\title{get_default_embedding}
\usage{
get_default_embedding(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\value{
The default embedding.
}
\description{
Get the default embedding for the given .loom.
}
