% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_dgem}
\alias{get_dgem}
\title{get_dgem}
\usage{
get_dgem(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\value{
The gene expression matrix (genes as rows, cells as columns).
}
\description{
Get the expression matrix for the given .loom.
}
