% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_dspace}
\alias{get_dspace}
\title{get_dspace}
\usage{
get_dspace(x)
}
\arguments{
\item{x}{The dataspace interface name. Either scalar or simple.}
}
\value{
The corresponding HDF5 dataspace interface object.
}
\description{
Get the HDF5 dataspace interface object given x.
}
