% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_dtype}
\alias{get_dtype}
\title{get_dtype}
\usage{
get_dtype(x)
}
\arguments{
\item{x}{An R object or string describing HDF5 datatype.}
}
\value{
The corresponding HDF5 data type.
}
\description{
Get HDF5 data types.
}
\seealso{
\link{hdf5r::h5types}
}
\author{
mojaveazure
}
