% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_embedding_by_name}
\alias{get_embedding_by_name}
\title{get_embedding_by_name}
\usage{
get_embedding_by_name(loom, embedding.name)
}
\arguments{
\item{loom}{The loom file handler.}

\item{embedding.name}{The name of the given embedding to retrieve}
}
\value{
The embedding with the given embedding.name.
}
\description{
Get the embedding with the given embedding.name in the given .loom.
}
