% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_embeddings}
\alias{get_embeddings}
\title{get_embeddings}
\usage{
get_embeddings(loom)
}
\arguments{
\item{loom}{.loom file name}
}
\value{
List with the embeddings (each of them as data.frame)
}
\description{
Get embeddings (e.g. t-SNE, UMAP) from the given loom file
}
\examples{
loomPath <-  file.path(system.file('extdata', package='SCopeLoomR'), "example.loom")
loom <- open_loom(loomPath)

embeddings <- get_embeddings(loom)
names(embeddings)
head(embeddings[[1]])
}
