% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R, R/loom_mine.R
\name{get_genes}
\alias{get_genes}
\alias{get_genes}
\title{get_genes}
\usage{
get_genes(loom, is.flybase.gn = F)

get_genes(loom, is.flybase.gn = F)
}
\arguments{
\item{loom}{The loom file handler.}

\item{is.flybase.gn}{Whether to retrieve the Flybase gene identifiers or not.}

\item{loom}{The loom file handler.}

\item{is.flybase.gn}{Whether to retrieve the Flybase gene identifiers or not.}
}
\description{
Get the gene names either symbols or Flybase gene identifiers.

Get the gene names either symbols or Flybase gene identifiers.
}
