% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_global_attr}
\alias{get_global_attr}
\title{get_global_attr}
\usage{
get_global_attr(loom, key)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The key of the global attribute to get value from.}
}
\description{
Get the global attribute with the given key.
}
