% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_global_md_clustering_by_id}
\alias{get_global_md_clustering_by_id}
\title{get_global_md_clustering_by_id}
\usage{
get_global_md_clustering_by_id(loom, clustering.id)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.id}{The ID of the clustering you want to get the field value from.}
}
\description{
Get the clustering object from the global MetaData attribute with the given clustering.id
}
