% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_global_md_clustering_info_by_name}
\alias{get_global_md_clustering_info_by_name}
\title{get_global_md_clustering_info_by_name}
\usage{
get_global_md_clustering_info_by_name(loom, clustering.name, field = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.name}{The name of the clustering you want to get the field value from.}

\item{field}{The field name to get the value from.}
}
\description{
Get the clustering object from the global MetaData attribute with the given clustering.name
}
