% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R, R/loom_mine.R
\name{get_global_meta_data}
\alias{get_global_meta_data}
\alias{get_global_meta_data}
\title{get_global_meta_data}
\usage{
get_global_meta_data(loom)

get_global_meta_data(loom)
}
\arguments{
\item{loom}{The loom file handler.}

\item{loom}{The loom file handler.}
}
\description{
Get the global MetaData attribute as a R object.

Get the global MetaData attribute as a R object.
}
