% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_next_embedding_id}
\alias{get_next_embedding_id}
\title{get_next_embedding_id}
\usage{
get_next_embedding_id(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Get the next embedding ID
}
