% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_regulons}
\alias{get_regulons}
\title{get_regulons}
\usage{
get_regulons(loom, tfAsName = TRUE, tfSep = "_")
}
\arguments{
\item{loom}{.loom file name}

\item{tfAsName}{Whether to return only the TF name (calls \code{SCENIC::getTF()}), or the regulon name as stored in the loom file.}

\item{tfSep}{Character used as separator for the TF name and extra values stored in the regulon name. To be passed to SCENIC::getTF()}
}
\value{
The regulons as incidence matrix or list.
}
\description{
Get regulons from the given loom file
}
\examples{
loomPath <-  file.path(system.file('extdata', package='SCopeLoomR'), "example.loom")
loom <- open_loom(loomPath)

regulons <- get_regulons(loom)
head(regulons)
}
