% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_regulonThresholds}
\alias{get_regulonThresholds}
\title{get_regulonThresholds}
\usage{
get_regulonThresholds(loom, onlySelected = TRUE)
}
\arguments{
\item{loom}{.loom file name}

\item{onlySelected}{Whether to return only the selected threshold for each regulon (as vector), 
or also the alternative thresholds calculated by AUCell (as list).}
}
\description{
Get the AUC thresholds for the regulons in the given loom file
}
\examples{
loomPath <-  file.path(system.file('extdata', package='SCopeLoomR'), "example.loom")
loom <- open_loom(loomPath)

thresholds <- get_regulonThresholds(loom)
head(thresholds)
}
