% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_row_attr_by_key}
\alias{get_row_attr_by_key}
\title{get_row_attr_by_key}
\usage{
get_row_attr_by_key(loom, key)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of the row attribute.}
}
\description{
Get the row attribute with the given key.
}
