% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{get_seurat_clustering_resolutions}
\alias{get_seurat_clustering_resolutions}
\title{get_seurat_clustering_resolutions}
\usage{
get_seurat_clustering_resolutions(seurat, prefix = NULL)
}
\arguments{
\item{seurat}{A Seurat object.}

\item{prefix}{Prefix of the clustering to extract the resolutions from (only for Seurat v3 objects).}
}
\description{
Get list of all computed clusterings resolutions in the given seurat object.
}
