% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R, R/loom_mine.R
\name{getDtype}
\alias{getDtype}
\alias{getDtype}
\title{getDtype}
\usage{
getDtype(x)

getDtype(x)
}
\arguments{
\item{x}{An R object or string describing HDF5 datatype}

\item{x}{An R object or string describing HDF5 datatype}
}
\value{
The corresponding HDF5 data type

The corresponding HDF5 data type
}
\description{
Get HDF5 data types

Get HDF5 data types
}
\seealso{
\link{hdf5r::h5types}

\link{hdf5r::h5types}
}
