% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{global_meta_data_exists}
\alias{global_meta_data_exists}
\title{global_meta_data_exists}
\usage{
global_meta_data_exists(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Check if global MetaData attribute exists for the given loom.
}
