% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{hdf5_utf8_encode}
\alias{hdf5_utf8_encode}
\title{hdf5_utf8_encode}
\usage{
hdf5_utf8_encode(value, dtype)
}
\arguments{
\item{value}{The value from which the dtype should be UTF8 encoded.}

\item{dtype}{The dtype of the given value.}
}
\description{
Open loom file and return a .loom file handler.
}
