% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{hierarchy_exists}
\alias{hierarchy_exists}
\title{hierarchy_exists}
\usage{
hierarchy_exists(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Check if hierarchy exists for the given loom.
}
