% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{init_global_md_embeddings}
\alias{init_global_md_embeddings}
\title{init_global_md_embeddings}
\usage{
init_global_md_embeddings(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Remove all embeddings from the given loom.
}
