% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{init_global_meta_data}
\alias{init_global_meta_data}
\title{init_global_meta_data}
\usage{
init_global_meta_data(loom, loom.spec.version)
}
\arguments{
\item{loom}{The loom file handler.}

\item{loom.spec.version}{The version of the loom specification. Choose on of: 2 or 3}
}
\description{
Initialize the global MetaData attribute.
}
