% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{is_base64_encoded}
\alias{is_base64_encoded}
\title{is_base64_encoded}
\usage{
is_base64_encoded(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
T is base64 encoded otherwise F.
}
\description{
Checks whether the given value is base64 encoded.
}
