% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{is_gzipped}
\alias{is_gzipped}
\title{is_gzipped}
\usage{
is_gzipped(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
T if gzip compressed otherwise F.
}
\description{
Checks whether the given value is gzip compressed.
}
