% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{is_json}
\alias{is_json}
\title{is_json}
\usage{
is_json(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
T if json string character otherwise F.
}
\description{
Checks whether the given value is a json character
}
