% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{is_order_conserved}
\alias{is_order_conserved}
\title{is_order_conserved}
\usage{
is_order_conserved(x, y)
}
\arguments{
\item{x}{A vector of elements}

\item{y}{Another vector of elements}
}
\description{
Check if order between elements of the given vectors x an y are the same
}
