% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{list_clusterings_names}
\alias{list_clusterings_names}
\title{list_clusterings_names}
\usage{
list_clusterings_names(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\value{
A list of all clustering names
}
\description{
Get a list of all clustering names in the given loom
}
