% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{lookup_all_global_attr}
\alias{lookup_all_global_attr}
\title{lookup_all_global_attr}
\usage{
lookup_all_global_attr(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
List all global attributes of the given .loom file handler.
}
