% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{open_loom}
\alias{open_loom}
\title{open_loom}
\usage{
open_loom(file.path, mode = "r")
}
\arguments{
\item{file.path}{The file path to the .loom file.}

\item{mode}{"r" for read/write (default). "r" for read-only.}
}
\value{
A loom file handler
}
\description{
Open loom file and return a .loom file handler.
}
