% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{remove_col_attr}
\alias{remove_col_attr}
\title{remove_col_attr}
\usage{
remove_col_attr(loom, key)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of column attribute to remove.}
}
\description{
Remove the column attribute with the given key.
}
