% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{remove_global_attr}
\alias{remove_global_attr}
\title{remove_global_attr}
\usage{
remove_global_attr(loom, key)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The key of the global attribute to remove.}
}
\description{
Remove the global attribute with the given key.
}
