% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{remove_row_attr}
\alias{remove_row_attr}
\title{remove_row_attr}
\usage{
remove_row_attr(loom, key)
}
\arguments{
\item{key}{The name of the row attribute.}
}
\description{
Remove the row attribute with the given key.
}
