% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{sce}
\alias{sce}
\title{Input dataset example}
\format{An object of class \code{SingleCellExperiment} with 50 rows and 100 columns.}
\usage{
sce
}
\description{
Small SingleCellExperiment object for examples
}
\details{
Documentation for the data
}
\keyword{data}
