% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{sub_clustering_exists}
\alias{sub_clustering_exists}
\title{sub_clustering_exists}
\usage{
sub_clustering_exists(parent.clustering.id)
}
\arguments{
\item{parent.clustering.id}{The clustering id of the parent}
}
\description{
Check whether a sub clustering of the given parent.clustering.id exists in the MetaData
}
