% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{sub_embedding_group_exists}
\alias{sub_embedding_group_exists}
\title{sub_embedding_group_exists}
\usage{
sub_embedding_group_exists(parent.clustering.id)
}
\arguments{
\item{parent.clustering.id}{The id of the parent clustering}
}
\description{
Check whether a sub embedding group of the given parent.clustering.id exists in the MetaData
}
