% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{tSNE_expr}
\alias{tSNE_expr}
\title{tSNE example}
\format{
An object of class \code{matrix} with 100 rows and 2 columns.
}
\usage{
tSNE_expr
}
\description{
t-SNE for the expression matrix
}
\keyword{data}
