% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{update_cluster_descriptions_by_cluster_annotation_mapping_df}
\alias{update_cluster_descriptions_by_cluster_annotation_mapping_df}
\title{update_cluster_descriptions_by_cluster_annotation_mapping_df}
\usage{
update_cluster_descriptions_by_cluster_annotation_mapping_df(loom,
  clustering.name, annotation.df = NULL,
  annotation.df.cluster.id.column.name = NULL,
  annotation.df.annotation.column.name = NULL, in.place = TRUE)
}
\arguments{
\item{loom}{The loom file handler.}

\item{clustering.name}{The name of the clustering.}

\item{annotation.df}{The data.frame containing the cluster annotatiion mapping}

\item{annotation.df.cluster.id.column.name}{The column name of the given annotation.df containing the cluster ID values}

\item{annotation.df.annotation.column.name}{The column name of the given annotation.df containing the annotation values.}

\item{in.place}{If true update the given loom, otherwise return the update global meta data.}
}
\description{
Update the cluster descriptions of the clusters of the given clustering_name
}
