% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{update_col_attr}
\alias{update_col_attr}
\title{update_col_attr}
\usage{
update_col_attr(loom, key, value)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of column attribute to update.}

\item{value}{The new value to update.}
}
\description{
Update the column attribute with the given key and with the given value.
}
