% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{update_global_attr}
\alias{update_global_attr}
\title{update_global_attr}
\usage{
update_global_attr(loom, key, value)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The key of the global attribute to update.}

\item{value}{The new value}
}
\description{
Update the global attribute with the given key and the given value.
}
