% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{update_global_meta_data}
\alias{update_global_meta_data}
\title{update_global_meta_data}
\usage{
update_global_meta_data(loom, meta.data.json)
}
\arguments{
\item{loom}{The loom file handler.}

\item{meta.data.json}{The meta data stored as a json string.}
}
\description{
Update the global MetaData attribute with the given meta.data.json.
}
