% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{update_row_attr}
\alias{update_row_attr}
\title{update_row_attr}
\usage{
update_row_attr(loom, key, value)
}
\arguments{
\item{key}{The name of the row attribute to remove.}

\item{value}{The new value to update.}
}
\description{
Update the row attribute with the given key and the given value.
}
