% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_clustering}
\alias{add_clustering}
\title{add_clustering}
\usage{
add_clustering(loom, group, name, clusters, is.default = F,
  annotation = NULL, annotation.cluster.id.cl = NULL,
  annotation.cluster.description.cl = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{group}{The for the given clustering group to which the given clusters have to be added}

\item{clusters}{A list of the the cluster id for each cell present in the matrix}

\item{is.default}{}

\item{annotation}{A data.frame with annotation for the clusters}

\item{annotation.cluster.id.cl}{The column name to use for the IDs of the clusters found by the given clustering group.}

\item{annotation.cluster.description.cl}{The column name to use for the description of the clusters found by the given clustering group.}
}
\description{
Add the given clusters in the given group column attribute and meta data related to the given clustering to the given .loom file handler.
}
