% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_col_attr}
\alias{add_col_attr}
\title{add_col_attr}
\usage{
add_col_attr(loom, key, value, dtype = NULL, as.md.annotation = F)
}
\arguments{
\item{loom}{The loom file handler.}

\item{key}{The name of the new added attribute.}

\item{value}{The value of the new added attribute.}

\item{as.md.annotation}{Whether to show this attribute in the compare datasets tab.}
}
\description{
Add a new column attribute to the given .loom object accessible by the given key and containing the given value.
}
