% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_global_md_clustering}
\alias{add_global_md_clustering}
\title{add_global_md_clustering}
\usage{
add_global_md_clustering(loom, group, name, clusters, annotation = NULL,
  annotation.cluster.id.cl = NULL, annotation.cluster.description.cl = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{group}{The name of the group of clusterings}

\item{name}{The name given to the given clustering.}

\item{clusters}{A list of the the cluster id for each cell present in the same order as in the columns of gene expression matrix}

\item{annotation.cluster.id.cl}{The column name to use for the IDs of the clusters found by the given clustering group.}

\item{annotation.cluster.description.cl}{The column name to use for the description of the clusters found by the given clustering group.}
}
\description{
Add the clustering annotation to the global MetaData attribute.
}
