% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_global_md_regulon_thresholds}
\alias{add_global_md_regulon_thresholds}
\title{add_global_md_regulon_thresholds}
\usage{
add_global_md_regulon_thresholds(loom, regulon.threshold.assignments,
  regulon.enrichment.table)
}
\arguments{
\item{loom}{The loom file handler.}

\item{regulon.threshold.assignments}{The automated regulon cell assignments object generated by SCENIC.}

\item{regulon.enrichment.table}{The regulon enrichment table generated by SCENIC.}
}
\description{
Add the regulon thresholds annotation to the global MetaData attribute.
}
