% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R, R/loom_mine.R
\name{add_scenic_regulons}
\alias{add_scenic_regulons}
\alias{add_scenic_regulons}
\title{add_scenic_regulons}
\usage{
add_scenic_regulons(loom, dgem, regulons)

add_scenic_regulons(loom, dgem, regulons)
}
\arguments{
\item{loom}{The loom file handler.}

\item{dgem}{A matrix of the gene expression with M genes as rows and N cells as columns.}

\item{regulons}{A list of list of the regulons and their target genes generated by SCENIC.}

\item{loom}{The loom file handler.}

\item{dgem}{A matrix of the gene expression with M genes as rows and N cells as columns.}

\item{regulons}{A list of list of the regulons and their target genes generated by SCENIC.}
}
\description{
Add the regulons with their target genes generated by SCENIC as a row attribute to the given .loom file handler.

Add the regulons with their target genes generated by SCENIC as a row attribute to the given .loom file handler.
}
