% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_seurat_clustering}
\alias{add_seurat_clustering}
\title{add_seurat_clustering}
\usage{
add_seurat_clustering(loom, seurat, seurat.markers.file.path.list = NULL,
  seurat.marker.metric.accessors = NULL, seurat.marker.metric.names = NULL,
  seurat.marker.metric.description = NULL,
  default.clustering.resolution = NULL, annotation = NULL,
  annotation.cluster.id.cn = NULL, annotation.cluster.description.cn = NULL)
}
\arguments{
\item{loom}{The loom file handler.}

\item{seurat}{The Seurat object}

\item{seurat.markers.file.path.list}{The named list of file paths to the markers saved in RDS format. The names should be the resolution id of the corresponding clustering (e.g.: res.2). Default is NULL.}

\item{default.clustering.resolution}{The clustering resolution (i.e.: res.2, ...) of the clustering that should be set as the default which an annotation can be set for.}

\item{annotation}{A data.frame with annotation for the clusters of the default clustering. Default is NULL.}

\item{annotation.cluster.id.cn}{The column name to use for the IDs of the clusters found by the given clustering group. Default is NULL.}

\item{annotation.cluster.description.cn}{The column name to use for the description of the clusters found by the given clustering group. Default is NULL.}
}
\description{
Add all the Seurat clusterings in the given seurat object to the given .loom file handler.
}
