% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{add_seurat_clustering}
\alias{add_seurat_clustering}
\title{add_seurat_clustering}
\usage{
add_seurat_clustering(loom, seurat, default.clustering.resolution,
  seurat.markers.file.path.list, annotation, annotation.cluster.id.cl,
  annotation.cluster.description.cl)
}
\arguments{
\item{loom}{The loom file handler.}

\item{seurat}{The Seurat object}

\item{default.clustering.resolution}{The clustering resolution (i.e.: res.2, ...) of the clustering that should be set as the default.}

\item{seurat.markers.file.path.list}{The named list of file paths to the markers saved in RDS format. The names should be the resolution id of the corresponding clustering (e.g.: res2.0).}

\item{annotation}{A data.frame with annotation for the clusters}

\item{annotation.cluster.id.cl}{The column name to use for the IDs of the clusters found by the given clustering group.}

\item{annotation.cluster.description.cl}{The column name to use for the description of the clusters found by the given clustering group.}
}
\description{
Add all the Seurat clusterings in the given seurat object to the given .loom file handler.
}
