% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{build_loom}
\alias{build_loom}
\title{build_loom}
\usage{
build_loom(file.name, title = NULL, genome = NULL, dgem, default.embedding,
  default.embedding.name, fbgn.gn.mapping.file.path = NULL,
  chunk.size = 1000, display.progress = T)
}
\arguments{
\item{file.name}{A string naming the .loom file to be generated.}

\item{dgem}{A matrix of the gene expression with M genes as rows and N cells as columns.}

\item{default.embedding}{A M-by-2 data.frame of the embedding (X and Y coordinates) of the cells.}

\item{fbgn.gn.mapping.file.path}{A N-by-2 data.frame containing the mapping between the Flybase gene and the gene symbol.}
}
\description{
build_loom
}
