% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{init_global_meta_data}
\alias{init_global_meta_data}
\title{init_global_meta_data}
\usage{
init_global_meta_data(loom)
}
\arguments{
\item{loom}{The loom file handler.}
}
\description{
Empty the global MetaData attribute.
}
